refactor : 미션 페이지 리팩토링
